rename: 팟 상세페이지 폴더명 수정
