+++ src/pages/PotDetail/PotDetail.style.ts
