+++ src/pages/PotDetail/PotDetail.tsx
